requirements.txt
--find-links=https://girder.github.io/large_image_wheels GDAL
# cartopy
geemap
geopandas
jupyter-server-proxy
keplergl
leafmap
localtileserver
nbserverproxy
owslib
palettable
plotly
streamlit=>1.18.1
streamlit-bokeh-events
streamlit-folium
streamlit-keplergl
tropycal
# git+https://github.com/giswqs/leafmap
# git+https://github.com/giswqs/geemap

